refactor: organize src into types/, utils/, and scrapers/ subdirectories
